Moving creating listener middleware to the persistence layer.
